requirements.txt
supabase>=1.0.0
requests>=2.31.0
psycopg[binary]>=3.1
sqlalchemy>=1.4
pgvector==0.4.1
python-dotenv>=1.0.0
sqlmodel

# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LangChain dependencies
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# Google Generative AI (for Gemini embeddings)
#google-generativeai>=0.3.0

# JWT token handling
python-jose[cryptography]==3.3.0